Uninstall fix and documentantion

+++ scripts/uninstall.sh
@@ -4,31 +4,39 @@ LIB_DIR="lib"
 INC_DIR="include"
 
 LIBS=()
-for LIBPATH in "$LIB_DIR"/lib*.so; do
+while IFS= read -r -d '' LIBPATH; do
     LIBS+=("$LIBPATH")
-done
+done < <(find "$LIB_DIR" -maxdepth 1 -type f -name 'lib*.so' -print0)
 
 HEADERS=()
-for HEADERPATH in "$INC_DIR"/*.h; do
+while IFS= read -r -d '' HEADERPATH; do
     HEADERS+=("$HEADERPATH")
-done
+done < <(find "$INC_DIR" -maxdepth 1 -type f -name '*.h' -print0)
 
 LIB_LOCATION="/usr/local/lib"
 INC_LOCATION="/usr/local/include"
 
 #Copy library files
 
-for HEADER  in "${HEADERS[@]}"; do
-    NAME=$(basename "$HEADER")
-    echo "Removing $INC_LOCATION/$NAME"
-    rm -f "$INC_LOCATION/$NAME"
-done
+if [ ${#HEADERS[@]} -gt 0 ]; then
+    for HEADER in "${HEADERS[@]}"; do
+        NAME=$(basename "$HEADER")
+        echo "Removing $INC_LOCATION/$NAME"
+        rm -f "$INC_LOCATION/$NAME"
+    done
+else
+    echo "No header files found in $INC_DIR"
+fi
 
-for LIB  in "${LIBS[@]}"; do
-    NAME=$(basename "$LIB")
-    echo "Removing $LIB_LOCATION/$NAME"
-    rm -f "$LIB_LOCATION/$NAME"
-done
+if [ ${#LIBS[@]} -gt 0 ]; then
+    for LIB in "${LIBS[@]}"; do
+        NAME=$(basename "$LIB")
+        echo "Removing $LIB_LOCATION/$NAME"
+        rm -f "$LIB_LOCATION/$NAME"
+    done
+else
+    echo "No library files found in $LIB_DIR"
+fi
 
 # Setup linker
